requirements.txt
Flask==3.0.0
gunicorn==21.2.0
APScheduler==3.10.4
pytrends==4.9.2
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.55.3
google-generativeai==0.8.5
httpx==0.27.0
feedparser==6.0.11
python-dotenv==1.0.0
piexif==1.1.3
Pillow==11.0.0
pydantic>=2.0.0
cryptography>=41.0.0
playwright>=1.40.0
playwright-stealth>=1.0.6
